cria listas de ataque, defesa e suporte

+++ src/cartas/Ataque.java
@@ -1,7 +1,7 @@
 package cartas;
 
 public class Ataque extends Carta {
-    public Ataque(String nome, String tipo, int poder, int custo, String descricao){
+    public Ataque(String nome, String tipo, double poder, double custo, String descricao){
         this.nome = nome;
         this.tipo = tipo;
         this.poder = poder;
